requirements.txt
azure-functions
tinyapi
feedparser
azure-storage-blob